requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.3
mysql-connector-python==9.0.0
python-dotenv==1.0.1